Data.Functor.Contravariant.Decide

Description

This module is only available if building with GHC 8.6 or later, or if the +contravariant cabal build flag is available.

Documentation

class Contravariant f => Decide f where

The contravariant analogue of Alt.

If one thinks of f a as a consumer of as, then decide allows one to handle the consumption of a value by choosing to handle it via exactly one of two independent consumers. It redirects the input completely into one of two consumers.

decide takes the "decision" method and the two potential consumers, and returns the wrapped/combined consumer.

Mathematically, a functor being an instance of Decide means that it is "semigroupoidal" with respect to the contravariant "either-based" Day convolution (data EitherDay f g a = forall b c. EitherDay (f b) (g c) (a -> Either b c)). That is, it is possible to define a function (f EitherDay f) a -> f a in a way that is associative.

Since: 5.3.6

Methods

decide :: (a -> Either b c) -> f b -> f c -> f a

Takes the "decision" method and the two potential consumers, and returns the wrapped/combined consumer.

gdecide :: (Generic1 f, Decide (Rep1 f)) => (a -> Either b c) -> f b -> f c -> f a

Generic decide. Caveats:

Will not compile if f is a sum type.

Will not compile if f contains fields that do not mention its type variable.

-XDeriveGeneric is not smart enough to make instances where the type variable appears in negative position.

Since: 5.3.8

decided :: Decide f => f b -> f c -> f (Either b c)

For decided x y, the resulting f (Either b c) will direct Lefts to be consumed by x, and Rights to be consumed by y.

Since: 5.3.6

gdecided :: (Generic1 f, Decide (Rep1 f)) => f b -> f c -> f (Either b c)

Generic decided. Caveats are the same as for gdecide.

Since: 5.3.8