Data.Functor.Contravariant.Divise

Description

This module is only available if building with GHC 8.6 or later, or if the +contravariant cabal build flag is available.

Documentation

class Contravariant f => Divise f where

The contravariant analogue of Apply; it is Divisible without conquer.

If one thinks of f a as a consumer of as, then divise allows one to handle the consumption of a value by splitting it between two consumers that consume separate parts of a.

divise takes the "splitting" method and the two sub-consumers, and returns the wrapped/combined consumer.

All instances of Divisible should be instances of Divise with divise = divide.

If a function is polymorphic over Divise f (as opposed to Divisible f), we can provide a stronger guarantee: namely, that any input consumed will be passed to at least one sub-consumer. With Divisible f, said input could potentially disappear into the void, as this is possible with conquer.

Mathematically, a functor being an instance of Divise means that it is "semigroupoidal" with respect to the contravariant (tupling) Day convolution. That is, it is possible to define a function (f Day f) a -> f a in a way that is associative.

Since: 5.3.6

Methods

divise :: (a -> (b, c)) -> f b -> f c -> f a

Takes a "splitting" method and the two sub-consumers, and returns the wrapped/combined consumer.

gdivise :: (Divise (Rep1 f), Generic1 f) => (a -> (b, c)) -> f b -> f c -> f a

Generic divise. Caveats:

Will not compile if f is a sum type.

Will not compile if f contains fields that do not mention its type variable.

-XDeriveGeneric is not smart enough to make instances where the type variable appears in negative position.

Since: 5.3.8

divised :: Divise f => f a -> f b -> f (a, b)

Combine a consumer of a with a consumer of b to get a consumer of (a, b).

divised = divise id

Since: 5.3.6

gdivised :: (Generic1 f, Divise (Rep1 f)) => f a -> f b -> f (a, b)

Generic divised. Caveats are the same as for gdivise.

Since: 5.3.8

newtype WrappedDivisible f a

Wrap a Divisible to be used as a member of Divise

Since: 5.3.6

Constructors

WrapDivisible

Fields

unwrapDivisible :: f a