Data.Functor.Extend

Description

Extendable Functors

There are two ways to define an Extend instance:

I. Provide definitions for extend satisfying this law:

 extended f . extended g = extended (f . extended g)

II. Alternately, you may choose to provide definitions for duplicate satisfying this law:

 duplicated . duplicated = fmap duplicated . duplicated

These are both equivalent to the statement that (->-) is associative

 (f ->- g) ->- h = f ->- (g ->- h)

You may of course, choose to define both duplicate and extend. In that case you must also satisfy these laws:

 extended f = fmap f . duplicated
 duplicated = extended id

These are the default definitions of extended and duplicated.

class Functor w => Extend w where

Methods

duplicated :: w a -> w (w a)

 duplicated = extended id
 fmap (fmap f) . duplicated = duplicated . fmap f

extended :: (w a -> b) -> w a -> w b

 extended f  = fmap f . duplicated