Data.Functor.Plus

Description

Documentation

class Alt f => Plus f where

Laws:

zero <!> m = m
m <!> zero = m

If extended to an Alternative then zero should equal empty.

Methods

zero :: f a

psum :: (Foldable t, Plus f) => t (f a) -> f a

The sum of a collection of actions, generalizing concat.

>>> psum [Just "Hello", Nothing, Just "World"]
Just "Hello"

Since: 5.3.6

gzero :: (Plus (Rep1 f), Generic1 f) => f a

Generic zero. Caveats:

Will not compile if f is a sum type.

Any types where the a does not appear must have a Monoid instance.

Since: 5.3.8

module Data.Functor.Alt