Data.Semifunctor

Documentation

class (Semigroupoid c, Semigroupoid d) => Semifunctor f c d | f c -> d, f d -> c where

Semifunctors map objects to objects, and arrows to arrows preserving connectivity as normal functors, but do not purport to preserve identity arrows. We apply them to semigroupoids, because those don't even claim to offer identity arrows!

Methods

semimap :: c a b -> d (f a) (f b)

data Bi p a where

Used to map a more traditional bifunctor into a semifunctor

Constructors

Bi :: p a b -> Bi p (a, b)

(#) :: a -> b -> Bi (,) (a, b)

semibimap :: Semifunctor p (Product l r) cod => l a b -> r c d -> cod (p (a, c)) (p (b, d))

semifirst :: (Semifunctor p (Product l r) cod, Ob r c) => l a b -> cod (p (a, c)) (p (b, c))

semisecond :: (Semifunctor p (Product l r) cod, Ob l a) => r b c -> cod (p (a, b)) (p (a, c))

first :: (Semifunctor p (Product l r) cod, Category r) => l a b -> cod (p (a, c)) (p (b, c))

second :: (Semifunctor p (Product l r) cod, Category l) => r b c -> cod (p (a, b)) (p (a, c))

data WrappedFunctor f a

Constructors

WrapFunctor

Fields

unwrapFunctor :: f a

data WrappedTraversable1 f a

Constructors

WrapTraversable1

Fields

unwrapTraversable1 :: f a

module Control.Category

module Data.Semigroupoid

module Data.Semigroupoid.Ob

module Data.Semigroupoid.Product