Data.Semigroup.Foldable

Description

Re-exports a subset of the Data.Foldable1 module along with some additional combinators that require Foldable1 constraints.

Documentation

class Foldable t => Foldable1 (t :: Type -> Type) where

Non-empty data structures that can be folded.

Since: base-4.18.0.0

Minimal complete definition

foldMap1 | foldrMap1

Methods

fold1 :: Semigroup m => t m -> m

Combine the elements of a structure using a semigroup. @since 4.18.0.0

foldMap1 :: Semigroup m => (a -> m) -> t a -> m

Map each element of the structure to a semigroup, and combine the results.

>>> foldMap1 Sum (1 :| [2, 3, 4])
Sum {getSum = 10}

Since: base-4.18.0.0

toNonEmpty :: t a -> NonEmpty a

List of elements of a structure, from left to right.

>>> toNonEmpty (Identity 2)
2 :| []

Since: base-4.18.0.0

intercalate1 :: (Foldable1 t, Semigroup m) => m -> t m -> m

Insert an m between each pair of t m.

>>> intercalate1 ", " $ "hello" :| ["how", "are", "you"]
"hello, how, are, you"

>>> intercalate1 ", " $ "hello" :| []
"hello"

>>> intercalate1 mempty $ "I" :| ["Am", "Fine", "You?"]
"IAmFineYou?"

Since: base-4.18.0.0

foldrM1 :: (Foldable1 t, Monad m) => (a -> a -> m a) -> t a -> m a

Monadic fold over the elements of a non-empty structure, associating to the right, i.e. from right to left.

Since: base-4.18.0.0

foldlM1 :: (Foldable1 t, Monad m) => (a -> a -> m a) -> t a -> m a

Monadic fold over the elements of a non-empty structure, associating to the left, i.e. from left to right.

Since: base-4.18.0.0

intercalateMap1 :: (Foldable1 t, Semigroup m) => m -> (a -> m) -> t a -> m

Insert m between each pair of m derived from a.

>>> intercalateMap1 " " show $ True :| [False, True]
"True False True"

>>> intercalateMap1 " " show $ True :| []
"True"

traverse1_ :: (Foldable1 t, Apply f) => (a -> f b) -> t a -> f ()

for1_ :: (Foldable1 t, Apply f) => t a -> (a -> f b) -> f ()

sequenceA1_ :: (Foldable1 t, Apply f) => t (f a) -> f ()

foldMapDefault1 :: (Foldable1 t, Monoid m) => (a -> m) -> t a -> m

Usable default for foldMap, but only if you define foldMap1 yourself

asum1 :: (Foldable1 t, Alt m) => t (m a) -> m a

gfold1 :: (Foldable1 (Rep1 t), Generic1 t, Semigroup m) => t m -> m

Generic fold1. Caveats:

Will not compile if t is an empty constructor.

Will not compile if t has some fields that don't mention a, for exmaple data Bar a = MkBar a Int

Since: 5.3.8

gfoldMap1 :: (Foldable1 (Rep1 t), Generic1 t, Semigroup m) => (a -> m) -> t a -> m

Generic foldMap1. Caveats are the same as for gfold1.

Since: 5.3.8

gtoNonEmpty :: (Foldable1 (Rep1 t), Generic1 t) => t a -> NonEmpty a

Generic toNonEmpty. Caveats are the same as for gfold1.

Since: 5.3.8