Data.Semigroup.Foldable.Class

Description

Deprecated: This module re-exports a limited subset of the class methods in the Foldable1 and Bifoldable1 classes, which are now located in the Data.Foldable1 and Data.Bifoldable1 modules in base-4.18. (On older versions of base, these can be found in the foldable1-classes-compat library.) Import from these modules instead.

Documentation

class Foldable t => Foldable1 (t :: Type -> Type) where

Non-empty data structures that can be folded.

Since: base-4.18.0.0

Minimal complete definition

foldMap1 | foldrMap1

Methods

fold1 :: Semigroup m => t m -> m

Combine the elements of a structure using a semigroup. @since 4.18.0.0

foldMap1 :: Semigroup m => (a -> m) -> t a -> m

Map each element of the structure to a semigroup, and combine the results.

>>> foldMap1 Sum (1 :| [2, 3, 4])
Sum {getSum = 10}

Since: base-4.18.0.0

toNonEmpty :: t a -> NonEmpty a

List of elements of a structure, from left to right.

>>> toNonEmpty (Identity 2)
2 :| []

Since: base-4.18.0.0

class Bifoldable t => Bifoldable1 (t :: Type -> Type -> Type) where

Minimal complete definition

bifoldMap1

Methods

bifold1 :: Semigroup m => t m m -> m

bifoldMap1 :: Semigroup m => (a -> m) -> (b -> m) -> t a b -> m