Data.Semigroup.Traversable

Description

Documentation

class (Foldable1 t, Traversable t) => Traversable1 t where

Minimal complete definition

traverse1 | sequence1

Methods

traverse1 :: Apply f => (a -> f b) -> t a -> f (t b)

sequence1 :: Apply f => t (f b) -> f (t b)

Defining Traversable1 instances

Defining Traversable1 instances for types with both Traversable1 and Traversable substructures can be done with traverse1Maybe, (<*.>), and (<.*>).

data Foo a = Foo (Maybe a) (Maybe a) a [a]
  deriving (Functor, Traversable, Foldable)
instance Traversable1 Foo where
  traverse1 f (Foo ma ma' a as) = Foo <$> traverseMaybe ma <*> traverseMaybe ma' <*.> f a <.*> traverseMaybe as
instance Foldable1 Foo where
  foldMap1 = foldMap1Default

traverse1Maybe :: (Traversable t, Apply f) => (a -> f b) -> t a -> MaybeApply f (t b)

Traverse a Traversable using Apply, getting the results back in a MaybeApply.

gtraverse1 :: (Traversable1 (Rep1 t), Apply f, Generic1 t) => (a -> f b) -> t a -> f (t b)

Generic traverse1. Caveats:

Will not compile if t is an empty constructor.

Will not compile if t has some fields that don't mention a, for exmaple data Bar a = MkBar a Int

Since: 5.3.8

gsequence1 :: (Traversable1 (Rep1 t), Apply f, Generic1 t) => t (f b) -> f (t b)

Generic sequence1. Caveats are the same for gtraverse1.

Since: 5.3.8

Default superclass instance helpers

foldMap1Default :: (Traversable1 f, Semigroup m) => (a -> m) -> f a -> m

Default implementation of foldMap1 given an implementation of Traversable1.