Semigroupoids.Do

Description

This module re-exports operators from Data.Functor.Apply and Data.Functor.Bind, but under the same names as their Applicative and Monad counterparts. This makes it convenient to use do-notation on a type that is a Bind but not a monad (or an Apply but not an Applicative with ApplicativeDo), either using the QualifiedDo extension or the more traditional RebindableSyntax.

{-# LANGUAGE ApplicativeDo #-}
{-# LANGUAGE QualifiedDo #-}

foo :: Apply f => f a -> f b -> f (a, b)
foo as bs = Semi.do
  a <- as
  b <- bs
  pure (a, b)

bar :: Bind m => (a -> b -> m c) -> m a -> m b -> m c
bar f as bs = Semi.do
  a <- as
  b <- bs
  f a b

Documentation

fmap :: Functor f => (a -> b) -> f a -> f b

fmap is used to apply a function of type (a -> b) to a value of type f a, where f is a functor, to produce a value of type f b. Note that for any type constructor with more than one parameter (e.g., Either), only the last type parameter can be modified with fmap (e.g., b in `Either a b`).

Some type constructors with two parameters or more have a Bifunctor instance that allows both the last and the penultimate parameters to be mapped over.

Examples

Expand

Convert from a Maybe Int to a Maybe String using show:

>>> fmap show Nothing
Nothing
>>> fmap show (Just 3)
Just "3"

Convert from an Either Int Int to an Either Int String using show:

>>> fmap show (Left 17)
Left 17
>>> fmap show (Right 17)
Right "17"

Double each element of a list:

>>> fmap (*2) [1,2,3]
[2,4,6]

Apply even to the second element of a pair:

>>> fmap even (2,2)
(2,True)

It may seem surprising that the function is only applied to the last element of the tuple compared to the list example above which applies it to every element in the list. To understand, remember that tuples are type constructors with multiple type parameters: a tuple of 3 elements (a,b,c) can also be written (,,) a b c and its Functor instance is defined for Functor ((,,) a b) (i.e., only the third parameter is free to be mapped over with fmap).

It explains why fmap can be used with tuples containing values of different types as in the following example:

>>> fmap even ("hello", 1.0, 4)
("hello",1.0,True)

(<*) :: Apply f => f a -> f b -> f a

Since: 5.3.6

(*>) :: Apply f => f a -> f b -> f b

Since: 5.3.6

(<*>) :: Apply f => f (a -> b) -> f a -> f b

Since: 5.3.6

(>>) :: Bind m => m a -> m b -> m b

Since: 5.3.6

(>>=) :: Bind m => m a -> (a -> m b) -> m b

Since: 5.3.6

join :: Bind m => m (m a) -> m a

pure :: Applicative f => a -> f a

Lift a value.

return :: Monad m => a -> m a

Inject a value into the monadic type.

fail :: Plus m => String -> m a

Important note

This ignores whatever String you give it. It is a bad idea to use fail as a form of labelled error; instead, it should only be defaulted to when a pattern match fails.

Since: 5.3.6