Data.Functor.Alt

Description

Documentation

class Functor f => Alt f where

Laws:

<!> is associative:             (a <!> b) <!> c = a <!> (b <!> c)
<$> left-distributes over <!>:  f <$> (a <!> b) = (f <$> a) <!> (f <$> b)

If extended to an Alternative then <!> should equal <|>.

Ideally, an instance of Alt also satisfies the "left distribution" law of MonadPlus with respect to <.>:

<.> right-distributes over <!>: (a <!> b) <.> c = (a <.> c) <!> (b <.> c)

IO, Either a, ExceptT e m and STM instead satisfy the "left catch" law:

pure a <!> b = pure a

Maybe and Identity satisfy both "left distribution" and "left catch".

These variations cannot be stated purely in terms of the dependencies of Alt.

When and if MonadPlus is successfully refactored, this class should also be refactored to remove these instances.

The right distributive law should extend in the cases where the a Bind or Monad is provided to yield variations of the right distributive law:

(m <!> n) >>- f = (m >>- f) <!> (m >>- f)
(m <!> n) >>= f = (m >>= f) <!> (m >>= f)

Minimal complete definition

(<!>)

Methods

(<!>) :: f a -> f a -> f a infixl 3

<|> without a required empty

some :: Applicative f => f a -> f [a]

many :: Applicative f => f a -> f [a]

optional :: (Alt f, Applicative f) => f a -> f (Maybe a)

One or none.

galt :: (Generic1 f, Alt (Rep1 f)) => f a -> f a -> f a

Generic (<!>). Caveats:

Will not compile if f is a sum type.

Any types where the a does not appear must have a Semigroup instance.

Since: 5.3.8

module Data.Functor.Apply