Data.Functor.Extend

Description

Extendable Functors

There are two ways to define an Extend instance:

I. Provide definitions for extended satisfying this law:

extended f . extended g = extended (f . extended g)

II. Alternately, you may choose to provide definitions for duplicated satisfying this law:

duplicated . duplicated = fmap duplicated . duplicated

You may of course, choose to define both duplicated and extended. In that case you must also satisfy these laws:

extended f = fmap f . duplicated
duplicated = extended id

These are the default definitions of extended and duplicated.

class Functor w => Extend w where

Minimal complete definition

duplicated | extended

Methods

duplicated :: w a -> w (w a)

duplicated = extended id
fmap (fmap f) . duplicated = duplicated . fmap f

extended :: (w a -> b) -> w a -> w b

extended f  = fmap f . duplicated

gduplicated :: (Extend (Rep1 w), Generic1 w) => w a -> w (w a)

Generic duplicated. Caveats:

Will not compile if w is a product type.

Will not compile if w contains fields where the type variable appears underneath the composition of type constructors (e.g., f (g a)).

Since: 5.3.8

gextended :: (Extend (Rep1 w), Generic1 w) => (w a -> b) -> w a -> w b

Generic extended. Caveats are the same as for gduplicated.

Since: 5.3.8