Data.Groupoid

Description

Documentation

class Semigroupoid k => Groupoid k where

semigroupoid with inverses. This technically should be a category with inverses, except we need to use Ob to define the valid objects for the category

Minimal complete definition

inv

Methods

inv :: k a b -> k b a