Data.Semigroupoid

Description

A semigroupoid satisfies all of the requirements to be a Category except for the existence of identity arrows.

Documentation

class Semigroupoid c where

Category sans id

Methods

o :: c j k -> c i j -> c i k

newtype WrappedCategory k a b

Constructors

WrapCategory

Fields

unwrapCategory :: k a b

newtype Semi m a b

Constructors

Semi

Fields

getSemi :: m