Data.Semigroupoid.Dual

Description

A semigroupoid satisfies all of the requirements to be a Category except for the existence of identity arrows.

Documentation

newtype Dual k a b

Constructors

Dual

Fields

getDual :: k b a